refactor and clean up the tplink map parser into it's own class
