+++ homeassistant/components/tplink/vacuum.py
@@ -4,12 +4,10 @@ from __future__ import annotations
 
 from base64 import b64decode
 from dataclasses import dataclass
-from typing import Any, Final, Literal, TypedDict
+from typing import Any, Final
 
 from kasa import Device, Module
 from kasa.smart.modules.clean import Clean, Status
-import lz4.block
-import numpy as np
 import voluptuous as vol
 
 from homeassistant.components.vacuum import (
@@ -31,6 +29,7 @@ from .entity import (
     TPLinkModuleEntityDescription,
     async_refresh_after,
 )
+from .map_parser import GetRoomsResponse, MapData, TpLinkMapParser
 
 # Coordinator is used to centralize the data updates
 # For actions the integration handles locking of concurrent device request
@@ -49,7 +48,21 @@ STATUS_TO_ACTIVITY = {
 }
 
 GET_ROOM_SERVICE_SCHEMA: Final = make_entity_service_schema(
-    {vol.Optional("map_id"): vol.Range()}
+    {vol.Optional("map_id"): int}
+)
+
+CLEAN_ROOMS_SERVICE_SCHEMA: Final = make_entity_service_schema(
+    {
+        vol.Required("map_id"): int,
+        vol.Required("room_ids"): [int],
+    }
+)
+
+CUSTOM_QUERY_SERVICE_SCHEMA: Final = make_entity_service_schema(
+    {
+        vol.Required("query"): str,
+        vol.Optional("params"): dict,
+    }
 )
 
 
@@ -123,6 +136,20 @@ async def async_setup_entry(
         supports_response=SupportsResponse.ONLY,
     )
 
+    platform.async_register_entity_service(
+        "clean_rooms",
+        CLEAN_ROOMS_SERVICE_SCHEMA,
+        TPLinkVacuumEntity.clean_rooms.__name__,
+        supports_response=SupportsResponse.NONE,
+    )
+
+    platform.async_register_entity_service(
+        "custom_query",
+        CUSTOM_QUERY_SERVICE_SCHEMA,
+        TPLinkVacuumEntity.custom_query.__name__,
+        supports_response=SupportsResponse.ONLY,
+    )
+
 
 class TPLinkVacuumEntity(CoordinatedTPLinkModuleEntity, StateVacuumEntity):
     """Representation of a tplink vacuum."""
@@ -178,6 +205,30 @@ class TPLinkVacuumEntity(CoordinatedTPLinkModuleEntity, StateVacuumEntity):
         """Set fan speed."""
         await self._vacuum_module.set_fan_speed_preset(fan_speed.capitalize())
 
+    async def custom_query(
+        self, query: str, params: dict | None = None
+    ) -> ServiceResponse:
+        """Execute a custom query."""
+        return await self._vacuum_module.call(query, params)
+
+    async def clean_rooms(self, map_id: int, room_ids: list[int]) -> None:
+        """Start cleaning the specified rooms."""
+        if self._vacuum_module.status is Status.Paused:
+            await self._vacuum_module.resume()
+            return
+
+        await self._vacuum_module.call(
+            "setSwitchClean",
+            {
+                "clean_mode": 3,
+                "clean_on": True,
+                "clean_order": True,
+                "force_clean": False,
+                "map_id": map_id,
+                "room_list": room_ids,
+            },
+        )
+
     async def async_locate(self, **kwargs: Any) -> None:
         """Locate the device."""
         await self._speaker_module.locate()
@@ -230,161 +281,3 @@ class TPLinkVacuumEntity(CoordinatedTPLinkModuleEntity, StateVacuumEntity):
         if self._vacuum_module.has_feature("fan_speed_preset"):
             self._attr_fan_speed = self._vacuum_module.fan_speed_preset.lower()
         return True
-
-
-type VacuumCoordinate = tuple[int, int, int]
-
-
-class GetRoomsResponse(TypedDict):
-    """Get rooms response."""
-
-    rooms: dict[int, str]
-    map_id: int
-
-
-class MapData(TypedDict):
-    """Map data."""
-
-    map_id: int
-    name: str
-    vac_coor: VacuumCoordinate
-    real_vac_coor: VacuumCoordinate
-    map_locked: bool
-    resolution: int
-    resolution_unit: str
-    width: int
-    height: int
-    origin_coor: VacuumCoordinate
-    real_origin_coor: VacuumCoordinate
-    pix_len: int
-    map_hash: str
-    pix_lz4len: str
-    map_data: str
-    area_list: list[MapDataAreaItem]
-
-
-class MapDataRoomItem(TypedDict):
-    """Map data room item."""
-
-    type: Literal["room"]
-    id: int
-    name: str
-    color: int
-    suction: int
-    cistern: int
-    clean_number: int
-    floor_texture: int
-    carpet_strategy: int
-
-
-class MapDataVirtualWallItem(TypedDict):
-    """Map data virtual wall item."""
-
-    type: Literal["virtual_wall"]
-    id: int
-    vertexs: list[list[int]]  # List of [x, y] pairs
-
-
-type MapDataAreaItem = MapDataRoomItem | MapDataVirtualWallItem
-
-room_map_cache: dict[
-    str, np.ndarray[tuple[int, int], np.dtype[np.unsignedinteger]]
-] = {}
-
-
-class TpLinkMapParser:
-    """Parse the map data."""
-
-    _map_data: MapData
-
-    def __init__(self, map_data: MapData) -> None:
-        """Initialize the map parser."""
-        self._map_data = map_data
-
-    def get_rooms(self) -> dict[int, str]:
-        """Parse the map data."""
-        rooms: dict[int, str] = {}
-        for room in self._map_data["area_list"]:
-            if room["type"] == "room":
-                rooms[room["id"]] = b64decode(room["name"]).decode("utf-8")
-
-        return rooms
-
-    def get_current_room(self) -> int | None:
-        """Get the current room."""
-        # we're assuming the vacuum coordinates are 2d
-        if (
-            self._map_data["real_vac_coor"] is None
-            or self._map_data["real_vac_coor"][2] != 0
-        ):
-            return None
-
-        rooms = self.get_rooms()
-        room_map = self._get_room_map(rooms)
-
-        if room_map is None:
-            return None
-
-        pixel_x = int(
-            (self._map_data["real_vac_coor"][0] - self._map_data["real_origin_coor"][0])
-            / self._map_data["resolution"]
-        )
-        pixel_y = int(
-            (self._map_data["real_vac_coor"][1] - self._map_data["real_origin_coor"][1])
-            / self._map_data["resolution"]
-        )
-
-        # Ensure within bounds
-        pixel_x = max(0, min(pixel_x, self._map_data["width"] - 1))
-        pixel_y = max(0, min(pixel_y, self._map_data["height"] - 1))
-
-        room_id: int | None = room_map[pixel_y, pixel_x]
-
-        if room_id not in rooms:
-            return None
-
-        return room_id
-
-    def _get_room_map(self, rooms: dict[int, str]) -> np.ndarray | None:
-        cached_array = room_map_cache.get(self._map_data["map_hash"])
-        if cached_array is not None:
-            return cached_array
-
-        map_array = self._decode_map_data()
-        if map_array is None:
-            return None
-
-        # Convert to room IDs
-        # Values 1-100 are typically used for room IDs
-        room_map = np.zeros(
-            (self._map_data["height"], self._map_data["width"]), dtype=np.uint8
-        )
-
-        for i in rooms:
-            # Find pixels with this room ID
-            room_map[map_array == i] = i
-
-        room_map_cache[self._map_data["map_hash"]] = room_map
-
-        return room_map
-
-    def _decode_map_data(
-        self,
-    ) -> np.ndarray | None:
-        """Decode the map data to get the pixel-by-pixel representation."""
-        if "map_data" not in self._map_data or not self._map_data["map_data"]:
-            return None
-
-        # Decode the base64 map data
-        compressed_data = b64decode(self._map_data["map_data"])
-        decompressed_data = lz4.block.decompress(
-            compressed_data, self._map_data["pix_len"]
-        )
-
-        # This gives us a 2D array where each pixel value indicates:
-        # - Values 1-7 (in your case): Room IDs (corresponds to area_list items with type "room")
-        # - 255: Cleanable space that's not in a specific room
-        # - Other values: Barriers, walls, or non-cleanable areas
-        return np.frombuffer(decompressed_data, dtype=np.uint8).reshape(
-            self._map_data["height"], self._map_data["width"]
-        )
